Refactor: move more things to editor.rs that belong there
